requirements.txt
pbr>=0.5.21,<1.0

greenlet>=0.3.2

eventlet>=0.13.0
Paste
PasteDeploy>=1.5.0
Routes>=1.12.3
WebOb>=1.2.3,<1.3

-f http://tarballs.openstack.org/oslo.config/oslo.config-1.2.0a3.tar.gz#egg=oslo.config-1.2.0a3
oslo.config>=1.2.0a3

babel
jsonschema>=1.3.0
six
iso8601
kombu
SQLAlchemy>=0.7.8,<=0.7.99
sqlalchemy-migrate>=0.7.2
